+++ assets/src/js/app/components/site-report-tile/site-report-tile.vue
@@ -16,24 +16,28 @@
                 slot="caption">
             {{ site.last_audit | format-date }}
         </span>
-        <report-detail :site-id="site.id"/>
+        <report-detail :report="report"/>
     </tile>
 </template>
 
 <script>
-    import withLatestReport from '../../containers/with-latest-report';
     import Btn from '../base/btn/btn';
     import ReportDetail from '../report-detail/report-detail';
     import ScheduledJobIndicator from '../scheduled-job-indicator/scheduled-job-indicator';
     import Tile from '../tile/tile';
 
     export default {
-        components: { ScheduledJobIndicator, ReportDetail: withLatestReport(ReportDetail), Btn, Tile },
+        components: { ScheduledJobIndicator, ReportDetail, Btn, Tile },
         props: {
             /** @type {Sites.SiteModel} */
             site: {
                 type: Object,
                 required: true,
+            },
+            /** @type {Reports.Report} */
+            report: {
+                type: Object,
+                default: null,
             },
         },
     };