+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 5.13
+
 **Features**
+
 0. [Split application into seperate projects (Core, Worker, UI) #47](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/47)
 0. [Remove webhooks logic #50](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/50)
 0. [Create list to see scheduled sites #35](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/35)
@@ -12,10 +14,13 @@
 0. [Make project active/disabled in settings menu of project #3](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/3)
 0. [Cleanup doc #63](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/63)
 0. [Add i18n #58](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/58)
+0. [Remove "HOC" and implement provider components #83](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/83)
 
 **Bugs**
+
 0. [Large JSON reports crash the app #75](https://github.com/lighthouse-dashboard/lighthouse-dashboard/issues/75)
 
+---
 
 # 5.12
 0. [Add logout button #22](https://github.com/faebeee/lighthouse-dashboard/issues/22)